add new tests Signed-off-by: Ishaan Mittal <ishaanmittal123@gmail.com>
